Create Binarysearch_recursion.cpp
The code can be divided into two parts: the binary search function and the main function.

The binarysearch function takes four parameters: arr (the sorted array to search), num (the number to search for), s (the starting index of the current search range), and e (the ending index of the current search range). It returns a boolean value indicating whether the number was found in the array.

Inside the binarysearch function, the mid index is calculated as the average of the starting and ending indices ((s + e) / 2).

If the starting index s becomes greater than the ending index e, it means the number is not present in the current search range, so the function returns false.

If the number at the mid index arr[mid] is greater than the target number num, it means the target number should be located in the left half of the array. In this case, the ending index e is updated to mid - 1 to narrow down the search range.

If the number at the mid index arr[mid] is less than the target number num, it means the target number should be located in the right half of the array. In this case, the starting index s is updated to mid + 1 to narrow down the search range.

If none of the above conditions match, it means the target number has been found in the array, and the function returns true.

Finally, the binarysearch function is recursively called with the updated values of s and e to perform the next iteration of the binary search.

In the main function, the starting index s and ending index e for the search range are initialized. The user is prompted to enter the number to be searched (num).

The binarysearch function is called with the array arr, the target number num, and the search range specified by s and e.

If the binarysearch function returns true, the message "number is present" is displayed. Otherwise, if it returns false, the message "number is not present" is displayed.