Create delete middle element of stack
The solve function is defined. It takes three parameters: inputStack, which represents the stack of integers, cnt to keep track of the count, and size to store the size of the stack.

The base case of the recursion is checked: if cnt (the count) is equal to size/2, it means we have reached the middle element(s). In this case, the top element of the stack is removed using inputStack.pop(), and the function returns.

If the base case is not met, the function continues to execute. It stores the top element of the stack in the variable num using inputStack.top(), and then removes the top element from the stack using inputStack.pop().

The solve function is recursively called again, incrementing the cnt by 1 and passing the updated count and the original size parameter.

After the recursive call returns, the original num is pushed back onto the stack using inputStack.push(num).

The deleteMiddle function is called from somewhere else, passing the original stack and its size (inputStack and N) as arguments.

The deleteMiddle function initializes a count variable cnt to 0.

The solve function is called from deleteMiddle, passing the stack, the count, and the size of the stack (inputStack, cnt, and N).

After the solve function completes, the middle element(s) of the stack will be deleted